Direct logrus output to slog
This allows starting to use log/slog, slowly migrating
from logrus.

This changes formatting, from
> FATA[0001] Error ...

(with the FATA red, and the [0001] part, very unintuitively,
a timestamp in seconds) to
> ERROR Error ...

when interactive, and from
> time="2026-08-14T00:24:59+02:00" level=fatal msg="Error..."

to
> 2026/08/17 21:42:08 ERROR Error ...

when noninteractive.

I think that is good enough, probably a bit of an improvement;
we can tweak the formatting later if necessary.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>